+++ packages/cody-server/lib/src/http/response.d.ts
@@ -1,3 +1,3 @@
 import { CodyResponse } from "@proophboard/cody-types";
-export declare type EmitResponse = (res: CodyResponse) => void;
-export declare type ReplyCallback<T = any> = (reply: T) => Promise<CodyResponse>;
+export type EmitResponse = (res: CodyResponse) => void;
+export type ReplyCallback<T = any> = (reply: T) => Promise<CodyResponse>;