+++ packages/cody-utils/lib/filesystem.d.ts
@@ -1,4 +1,4 @@
-import * as fs from 'fs';
+import { MakeDirectoryOptions, WriteFileOptions } from 'fs';
 import { CodyResponse } from "@proophboard/cody-types";
-export declare const mkdirIfNotExistsSync: (path: string, options?: string | number | fs.MakeDirectoryOptions | null | undefined) => CodyResponse | null;
-export declare const writeFileSync: (file: string, content: string, options?: fs.WriteFileOptions | undefined) => CodyResponse | null;
+export declare const mkdirIfNotExistsSync: (path: string, options?: number | string | MakeDirectoryOptions | null) => CodyResponse | null;
+export declare const writeFileSync: (file: string, content: string, options?: WriteFileOptions) => CodyResponse | null;