+++ packages/cody-utils/lib/metadata.d.ts
@@ -1,3 +1,3 @@
 import { CodyResponse, Node } from "@proophboard/cody-types";
-export declare const parseJsonMetadata: <T>(node: Node) => T | CodyResponse;
+export declare const parseJsonMetadata: <T>(node: Node) => CodyResponse | T;
 export declare const getStringMetadata: (node: Node) => string | CodyResponse;