+++ packages/schema-to-typescript/lib/jsonschema.d.ts
@@ -22,18 +22,15 @@ export declare const isRootNamespace: (ref: string) => boolean;
 export interface ShorthandObject {
     [property: string]: ShorthandObject | string;
 }
-export declare const convertShorthandObjectToJsonSchema: (shorthand: ShorthandObject, namespace?: string | undefined) => JSONSchema | CodyResponse;
+export declare const convertShorthandObjectToJsonSchema: (shorthand: ShorthandObject, namespace?: string) => JSONSchema | CodyResponse;
 export declare const convertShorthandStringToJsonSchema: (shorthand: string, namespace: string) => JSONSchema | CodyResponse;
-export declare const parseShorthandValidation: (validation: string) => [
-    string,
-    string | number | boolean | {
-        $data: string;
-    }
-] | CodyResponse;
+export declare const parseShorthandValidation: (validation: string) => [string, string | number | boolean | {
+    $data: string;
+}] | CodyResponse;
 export declare const dereferenceSchema: (schema: JSONSchema, defs: SchemaDefinitions) => Promise<JSONSchema>;
 export declare const relativeImportPath: (sourcePath: string, refSource: string) => string;
 export declare const isArrayType: (def: string, defs: SchemaDefinitions) => boolean;
 export declare const isStateType: (def: string, defs: SchemaDefinitions) => Promise<boolean>;
-export declare const mapPropertiesToTitles: (schema: JSONSchema, property?: string | undefined) => JSONSchema;
+export declare const mapPropertiesToTitles: (schema: JSONSchema, property?: string) => JSONSchema;
 export declare const definitionsContainReference: (ref: string, defs: SchemaDefinitions) => boolean;
 export declare const getSchemaFromDefinitions: (def: string, defs: SchemaDefinitions) => JSONSchema | null;