+++ packages/schema-to-typescript/lib/jsonschema.js
@@ -1,7 +1,11 @@
 "use strict";
 var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
     if (k2 === undefined) k2 = k;
-    Object.defineProperty(o, k2, { enumerable: true, get: function() { return m[k]; } });
+    var desc = Object.getOwnPropertyDescriptor(m, k);
+    if (!desc || ("get" in desc ? !m.__esModule : desc.writable || desc.configurable)) {
+      desc = { enumerable: true, get: function() { return m[k]; } };
+    }
+    Object.defineProperty(o, k2, desc);
 }) : (function(o, m, k, k2) {
     if (k2 === undefined) k2 = k;
     o[k2] = m[k];
@@ -369,10 +373,12 @@ const convertShorthandObjectToJsonSchema = (shorthand, namespace) => {
                 }
             }
             let itemsShorthandSchema = shorthand[schemaProperty];
-            if (itemsShorthandSchema.slice(-2) !== '[]') {
+            if (typeof itemsShorthandSchema === "string" && itemsShorthandSchema.slice(-2) !== '[]') {
                 itemsShorthandSchema = itemsShorthandSchema + '[]';
             }
-            const arraySchema = (0, exports.convertShorthandStringToJsonSchema)(itemsShorthandSchema, namespace);
+            const arraySchema = typeof itemsShorthandSchema === "string"
+                ? (0, exports.convertShorthandStringToJsonSchema)(itemsShorthandSchema, namespace)
+                : { type: "array", items: (0, exports.convertShorthandObjectToJsonSchema)(itemsShorthandSchema) };
             if (!(0, lib_1.isCodyError)(arraySchema) && Object.keys(shorthand).includes('$title')) {
                 arraySchema.title = shorthand['$title'];
             }