+++ packages/schema-to-typescript/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @proophboard/schema-to-typescript
 
+## 0.4.0
+
+### Minor Changes
+
+- Updated dependencies and shorthand schema can now contain nested objs in array
+
+### Patch Changes
+
+- Updated dependencies
+  - @proophboard/cody-types@0.5.0
+
 ## 0.3.2
 
 ### Patch Changes